+++ src/main/java/kr/co/pawong/pwbe/shelter/application/port/in/dto/ShelterCreate.java
@@ -1,7 +1,7 @@
 package kr.co.pawong.pwbe.shelter.application.port.in.dto;
 
 import kr.co.pawong.pwbe.shelter.enums.DivisionNm;
-import kr.co.pawong.pwbe.shelter.adapter.out.persistence.entity.ShelterEntity;
+import kr.co.pawong.pwbe.shelter.adapter.out.persistence.jpa.entity.ShelterEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
